+++ src/main/java/com/jack/priceservice/service/PriceService.java
@@ -1,6 +1,9 @@
 package com.jack.priceservice.service;
 
+import java.math.BigDecimal;
+
 public interface PriceService {
-    void setPrice(int price);
-    int getPrice();
+    void setPrice(BigDecimal price);
+
+    BigDecimal getPrice();
 }
